refactor(/strategy): NoLimitedDistanceMoveStrategy 추상 클래스로 분리
- 거리 제한 없는 moveStrategy 들을 위한 추상 클래스 분리
