+++ src/test/java/chess/domain/piece/PieceColorTest.java
@@ -3,7 +3,7 @@ package chess.domain.piece;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 class PieceColorTest {
